requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blessings @ file:///tmp/build/80754af9/blessings_1614076441300/work
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.5
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.3
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.0.0
faiss-gpu==1.7.2
fastapi==0.98.0
fastjsonschema==2.16.2
fishervector==0.1
flatbuffers==23.5.26
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2023.6.0
gpustat @ file:///tmp/build/80754af9/gpustat_1621432778913/work
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
inquirer==3.1.3
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lightning==2.0.4
lightning-cloud==0.5.37
lightning-utilities==0.8.0
llvmlite==0.39.1
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mpmath==1.2.1
multidict==6.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.12
numba==0.56.3
numpy==1.23.3
nvidia-ml-py==11.495.46
nvidia-ml-py3 @ file:///tmp/build/80754af9/nvidia-ml_1621428042568/work
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.13.0
PyJWT==2.7.0
pynndescent==0.5.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.4
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.0
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
readchar==4.0.5
requests==2.28.1
rich==13.4.2
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
tensorboardX==2.6
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.12
uvicorn==0.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.3
yarl==1.9.2
zipp==3.8.1
